feat. postPage 컴포넌트화 및 header style 세부조정 #10
